requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.7                          # via django
attrs==19.3.0                           # via pytest
certifi==2020.4.5.1                     # via requests
chardet==3.0.4                          # via requests
django==3.0.5                           # via -r requirements.in, djangorestframework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.6.0    # via -r requirements.in
djangorestframework==3.11.0             # via -r requirements.in, djangorestframework-simplejwt
idna==2.9                               # via requests
more-itertools==8.2.0                   # via pytest
packaging==20.3                         # via pytest
pluggy==0.13.1                          # via pytest
py==1.8.1                               # via pytest
pyjwt==2.0.0                            # via djangorestframework-simplejwt
pyparsing==2.4.7                        # via packaging
pytest-django==3.9.0                    # via -r requirements.in
pytest==5.4.1                           # via pytest-django
pytz==2019.3                            # via django
requests==2.23.0                        # via -r requirements.in
six==1.14.0                             # via packaging
sqlparse==0.3.1                         # via django
urllib3==1.25.9                         # via requests
wcwidth==0.1.9                          # via pytest
